[Refactor] ScaledStage 구조/타입 정리 & 리팩토링
